requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.1.2
django-haystack==2.8.1
django-mdeditor==0.1.9
gunicorn==19.9.0
idna==2.7
jieba==0.39
Markdown==3.0.1
Pillow==5.3.0
pycparser==2.19
Pygments==2.2.0
PyMySQL==0.9.2
python-slugify==1.2.6
pytz==2018.6
six==1.11.0
Unidecode==1.0.22
